arduino subscriber, publisher seems to be working

+++ RJJE_Arm/rjje_arm_ws/src/rjje_arm/arduino_files/servo_control/servo_control.h
@@ -18,7 +18,7 @@
 #define RESET 0xFF
 #define OKAY 0x0F    //this is a crappy design, but we can use this since we're just sending two decimal places
 
-inline int sign(double i){
+inline int sign(float i){
   if (i == 0) return 0; 
   else if (i < 0) return -1; 
   else return 1;
@@ -29,14 +29,14 @@ struct Motor{
     // Below all angles follow the right-hand convention 
 
     // In my setup, all motors have the neutral position at 90
-    double last_angle_ = 90;
+    float last_angle_ = 90;
     char offset_=0;   //offset should is added on commanded_angle
     // Therefore we need to "flip" the angle about the neutral position if necessary.
     bool is_claw_ = false;
     bool flip_rotation_ = false;   
 
-    bool set_angle(const double& commanded_angle, const short& channel_id, const Adafruit_PWMServoDriver& pwm){
-        double real_angle = get_real_angle(commanded_angle);
+    bool set_angle(const float& commanded_angle, const short& channel_id, const Adafruit_PWMServoDriver& pwm){
+        float real_angle = get_real_angle(commanded_angle);
         if (real_angle != -1){
           int pulselength = map(real_angle, 0, 180, SERVO_MIN, SERVO_MAX);
           pwm.setPWM(channel_id, 0, pulselength);
@@ -54,9 +54,9 @@ struct Motor{
     * @return: angle to be executed on servo after neccesary conversion. 
               -1 if the commanded_angle exceeds angle limits
     */
-    double get_real_angle (double commanded_angle){
+    float get_real_angle (float commanded_angle){
        if (is_claw_){
-           double rotation_angle = commanded_angle/2;
+           float rotation_angle = commanded_angle/2;
            //we're hard coding this because we're running out of memory
            if (0 <= rotation_angle && rotation_angle <= 90){
               return 90-rotation_angle + offset_;    //offset due to mislignment of rudder and finger 
@@ -65,7 +65,7 @@ struct Motor{
        }
        else{
          commanded_angle += offset_; 
-         double real_angle = (flip_rotation_) ? 180 - commanded_angle : commanded_angle;
+         float real_angle = (flip_rotation_) ? 180 - commanded_angle : commanded_angle;
          if (0 <= real_angle && real_angle <= 180){
            real_angle = last_angle_ + 0.5*(real_angle - last_angle_) ; 
            last_angle_ = real_angle;
@@ -78,7 +78,7 @@ struct Motor{
     }
 }; 
 
-inline void back_to_neutral(Motor* const motors, const Adafruit_PWMServoDriver& pwm, double* commanded_angles){
+inline void back_to_neutral(Motor* const motors, const Adafruit_PWMServoDriver& pwm, float* commanded_angles){
     commanded_angles[0] = 90; 
     commanded_angles[1] = 90; 
     commanded_angles[2] = 90; 
@@ -90,8 +90,8 @@ inline void back_to_neutral(Motor* const motors, const Adafruit_PWMServoDriver& 
     }
 }
 
-void test_arm(double* const commanded_angles, Motor* const motors, const double& step_angle, const Adafruit_PWMServoDriver& pwm){
-    static double degrees[6] = {90, 90, 90, 30, 90, 120};
+void test_arm(float* const commanded_angles, Motor* const motors, const float& step_angle, const Adafruit_PWMServoDriver& pwm){
+    static float degrees[6] = {90, 90, 90, 30, 90, 120};
     for (unsigned int channel_id = 0; channel_id < 6; ++channel_id){
         commanded_angles[channel_id] += sign(degrees[channel_id] - commanded_angles[channel_id]) * step_angle;
         motors[channel_id].set_angle(commanded_angles[channel_id], channel_id, pwm);
